+++ Library.Services/Implementations/UserUseCases/DeleteUserUseCase.cs
@@ -1,5 +1,6 @@
-﻿using Library.Application.Interfaces.Repositories;
-using Library.Domain.Interfaces.UserUseCases;
+﻿using Library.Application.Exceptions;
+using Library.Application.Interfaces.Repositories;
+using Library.Application.Interfaces.UserUseCases;
 
 namespace Library.Application.Implementations.UserUseCases {
     public class DeleteUserUseCase: IDeleteUserUseCase {
@@ -11,6 +12,9 @@ namespace Library.Application.Implementations.UserUseCases {
 
         public async Task Execute( Guid userId ) {
             var user = await _unit.userRepository.GetAsync( userId );
+            if (user == null) {
+                throw new NotFoundException( "There is no such user!" );
+            }
             await _unit.userRepository.DeleteAsync( user );
             await _unit.Save();
         }
